backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
scikit-learn==1.6.1
firebase-admin==6.7.0
requests==2.32.3
anthropic==0.52.0
gunicorn==23.0.0
